restructure tests according to best practices
