+++ libs/partners/perplexity/tests/unit_tests/test_secrets.py
@@ -5,5 +5,4 @@ def test_chat_perplexity_secrets() -> None:
     model = ChatPerplexity(
         model="llama-3.1-sonar-small-128k-online", pplx_api_key="foo"
     )
-    model_str = str(model)
-    assert "foo" not in model_str
+    assert "foo" not in str(model)
